Separate main nav items array from main controller
